resource sample pass

+++ rbc/world/include/rbc_world/resource.h
@@ -2,7 +2,7 @@
 #include "rbc_world/generated/resource_type.hpp"
 
 namespace rbc {
-using ResourceID = uint32_t;
+using ResourceID = uint64_t;
 constexpr ResourceID INVALID_RESOURCE = 0;
 
 // 资源元数据
